% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.select}
\alias{coef.select}
\title{PRESCIENCE estimates for coefficients}
\usage{
\method{coef}{select}(object, ...)
}
\arguments{
\item{object}{an object of class \code{select}.}

\item{...}{additional parameters.}
}
\value{
the input object is returned silently.
}
\description{
\code{coef.select} produces the estimated coefficients of the PRESCIENCE
produced by the \code{\link{select}} function.
}
\examples{
results <- select(auto ~ dcost + cars + dovtt + divtt,
data = transportation, nfoc = 1, q = 1, bound = 10)

coef(results)
}
\author{
Yankang (Bennie) Chen <yankang.chen@columbia.edu>
}
