% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prescience.R
\docType{package}
\name{prescience-package}
\alias{prescience-package}
\alias{prescience}
\title{Functions for the Approximate Best Subset Maximum Binary Prediction Rule (PRESCIENCE).}
\description{
\code{prescience} provides tools for best subset binary selection.
}
