% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Approximate Best Subset Maximum Binary Prediction Rule (PRESCIENCE)}
\usage{
select(formula, data, nfoc, q, bound, beta0 = 1, tol = NULL,
  warmstart = TRUE, tau = 1.5, mio = 1, tlim = 86400)
}
\arguments{
\item{formula}{an object of class formula of the format: (binary dependent variable) ~ (normalized focus variable) + (remaining focus variables) + (auxiliary variables).}

\item{data}{a data frame containing the variables in the model.}

\item{nfoc}{integer. The number of focus variable(s) excluding the intercept.}

\item{q}{integer. The cardinality constraint for the covariate selection.}

\item{bound}{numeric. The maximum absolute value of the bounds for all variables.}

\item{beta0}{integer. The coefficient taking value either 1 or -1 to normalize the scale for the first focus variable.}

\item{tol}{numeric. Tolerance level. If \code{NULL}, use the default tolerance level.}

\item{warmstart}{logical. If \code{TRUE}, use the warm start strategy.}

\item{tau}{the tuning parameter for enlarging the estimated bounds.}

\item{mio}{integer. 1 for MIO method 1 and 2 for method 2 in the paper.}

\item{tlim}{time limit (in seconds) specified for the MIO solver.}
}
\value{
a list with 7 elements:
\item{tolerance}{tolerance level}
\item{status}{optimization status}
\item{score}{Gurobi score}
\item{gap}{the MIO optimization gap value in case of early termination (0 if optimal solution is found within the time limit)}
\item{rtime}{time used by the MIO solver in the estimation procedure}
\item{ncount}{the number of Branch-and-bound nodes}
\item{bhat}{maximum score estimates for coefficients}
}
\description{
\code{select} employs the Approximate Best Subset Maximum Binary Prediction Rule
(PRESCIENCE) on the given variable selection problem.
}
\examples{
results <- select(auto ~ dcost + cars + dovtt + divtt,
data = transportation, nfoc = 1, q = 1, bound = 10)

summary(results)
coef(results)
}
\references{
\emph{Best Subset Binary Prediction} by Le-Yu Chen and Sokbae Lee (2018).
\url{https://arxiv.org/abs/1610.02738}
}
\author{
Yankang (Bennie) Chen <yankang.chen@columbia.edu>
}
